+++ pkg/api/results/network.go
@@ -39,9 +39,9 @@ type NetworkNATInfo struct {
 
 // NetworkInspect is the structured response for network inspection.
 type NetworkInspect struct {
-	Network NetworkItemInfo    `json:"network"`
-	Status  NetworkStatusInfo  `json:"status"`
-	NAT     NetworkNATInfo     `json:"nat"`
-	Leases  []NetworkLease     `json:"leases"`
+	Network NetworkItemInfo       `json:"network"`
+	Status  NetworkStatusInfo     `json:"status"`
+	NAT     NetworkNATInfo        `json:"nat"`
+	Leases  []NetworkLease        `json:"leases"`
 	Rules   []*model.FirewallRule `json:"rules,omitempty"`
 }